pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graph-issuer"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "python_async_framework @ git+https://github.com/michael-tran-91/python-async-framework.git@github.v1.0.3",
    "pyjwt",
    "cryptography"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]